+++ Coimbra.Services.PlayerLoopEvents/IPlayerLoopService.cs
@@ -3,7 +3,7 @@
 namespace Coimbra.Services.PlayerLoopEvents
 {
     /// <summary>
-    /// Responsible for any <see cref="IPlayerLoopEvent"/>, which includes <see cref="UpdateEvent"/>, <see cref="LateUpdateEvent"/>, <see cref="FixedUpdateEvent"/>, and each <see cref="PlayerLoopTimingEvents"/>.
+    /// Responsible for any <see cref="IPlayerLoopEvent"/>, which includes <see cref="UpdateEvent"/>, <see cref="LateUpdateEvent"/>, <see cref="FixedUpdateEvent"/>, and each <see cref="PlayerLoopInjectedTimings"/>.
     /// </summary>
     [RequiredService]
     public interface IPlayerLoopService : IService
@@ -11,15 +11,13 @@ namespace Coimbra.Services.PlayerLoopEvents
         /// <summary>
         /// Gets or sets the list of events currently being fired.
         /// </summary>
-        PlayerLoopTimingEvents CurrentTimings { get; set; }
+        PlayerLoopInjectedTimings CurrentInjectedTimings { get; set; }
 
         /// <inheritdoc cref="IEventService.AddListener{TEvent}"/>
         EventHandle AddListener(SerializableType<IPlayerLoopEvent> eventType, PlayerLoopEventHandler eventHandler);
 
-        /// <summary>
-        /// Removes all listeners from the specified type.
-        /// </summary>
-        void RemoveAllListeners<T>()
+        /// <inheritdoc cref="IEventService.RemoveAllListeners{TEvent}"/>
+        bool RemoveAllListeners<T>()
             where T : IPlayerLoopEvent;
     }
 }