+++ CHANGELOG.md
@@ -2,15 +2,18 @@
 
 ## [10.10.4] - UNRELEASED
 
+- Added `EventSettings` in the `Project Settings` window.
 - Added `Actor.OnStarting` to make API more consistent.
 - Added asserting to `Actor.OnDestroying` to avoid unexpected errors.
 - Added `IServiceFactory.ShouldSetService` property that defaults to true to keep previous behaviour.
 - Added more in-depth information for many existing APIs to make their usage clearer.
+- Changed `PlayerLoopTimingEvents` to `PlayerLoopInjectedTimings`.
 - Changed documentation structure to improve its readability.
 - Changed `DebugOnly` to `DebugOnlyComponent` as it is an `ActorComponent` to make API more consistent.
 - Changed `StartupSceneManager` to `EditorStartupSceneSettings` to make API more consistent.
 - Deprecated `StartListener` in favor of new `Actor.OnStarting` event. 
 - Fixed runtime issues with `ITimerService`.
+- Removed `COIMBRA_EVENTS_DISABLE_SAFETY_CHECKS` scripting define symbol in favor of new `EventSettings` options.
 - Removed `UPMAuthenticator` as it never really worked as it was expected.
 - Removed `ServiceLocator.GetFactory` overloads as it shouldn't be possible to retrieve the factories after set.
 - Removed `IntString` class as `int.ToString()` is already optimized.